requirements.txt
pathway[xpack-llm]
groq
fastembed
python-dotenv
requests